fix(review): enable @mention autocomplete in inline draft comments
Inline draft comments passed candidates={[]} to MentionTextarea, so the
@mention menu never opened there (only the activity/reply composers had
it). Extract collectMentionCandidates into a shared util, and thread a
candidate set (loaded comment authors + the PR author) from DiffView
through useDraftZones → DraftZoneList → DraftZone into the editor.
